chore(refactor): separate controllers from routers
- moves request controllers into own file

[Finishes #157840335]
